refactor pages into view components
