+++ src/pages/Account/index.tsx
@@ -1,0 +1,3 @@
+import AccountView from './View'
+
+export default AccountView
