+++ src/pages/ForRestaurants/index.tsx
@@ -1,0 +1,3 @@
+import ForRestaurantsView from './View'
+
+export default ForRestaurantsView
