+++ src/pages/Home/index.tsx
@@ -1,0 +1,3 @@
+import HomeView from './View'
+
+export default HomeView
