+++ src/pages/Restaurants/styles.module.css
@@ -1,0 +1,6 @@
+.container {
+  background-color: #EDE5CC;
+  min-height: 100vh;
+  padding-top: 3rem;
+  padding-bottom: 3rem;
+}
