+++ src/pages/Subscriptions/index.tsx
@@ -1,0 +1,3 @@
+import SubscriptionsView from './View'
+
+export default SubscriptionsView
